src/chatbot_thread1/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
chromadb>=0.4.0
numpy>=1.24.0
pydantic>=2.0.0

# Optional: Cho API endpoints
fastapi>=0.104.0
uvicorn>=0.24.0
flask>=3.0.0
flask-cors>=4.0.0
